AI EMAIL GENERATOR/requirements.txt
streamlit
transformers
google-generativeai
python-dotenv
python-multipart
uvicorn
fastapi
beautifulsoup4
requests
streamlit-option-menu
pyttsx3
googletrans